Error Handling.
Made a few adjustments to some if statements. I deleted the if statements
that was giving a problem and just went for the original option which
works.

I have an idea of why the if statement was not working which I beleive
it has something to do with the way variables are used. I will be looking
into it further to see if I can make it work.