Error Handling
An error occurred while archiving a file with 'tar' which kept the
script from running. The error in question was "File changed as we read it".

After troubleshooting for a while I find out that the file I was trying to archive
was being used by another program. I stopped the program and it ran smoothly.

While I was doing that I noticed something that got my attention. The 'find'
command's output was displaying the name of the directory and the files inside.
That output would then be piped to 'xargs' and run the 'tar' command to archive
the file. Even if the 'find' command were to exclude a file from that directory
it would still appear in the archived file. Because the output of 'find' displays
the absolute path of the directory to be archived. I'll need to find a way for
'find' to display only the files to be piped through to 'tar'.