Edited a Couple of Files.
I made a few modifications on the configure file so when run
it checks for certain commands that are installed. If they are
not installed it stops asks you to install them manually. Then,
after it checks for those commands it copies the main script
to /usr/local/bin directory and changes the permissions so it
can be executed without a problem.