Made some changes.
I made a few changes to some of the variables used to compress the files.
I also made other changes for error handling. I haven't yet tested the
function for the error handling. I still need to make other changes to the
script to see if the error handling function works.