Error Handling Function
I am working on a better error handling mechanic for the script.
The function, after it detects an error, display the error on standard
output then pipes it through the 'tee' command and writes the same
message to a file. The problem iswhen the error message is being piped
through 'tee' the color escape sequences format appears in the file.