Modified the Script.
I created a few more functions that will make the script more readable
and run smoothly. I also made a new menu that can help in choosing the
type of backup to use. I also wrote notes on the things I will use and
need for the script to function as expected. I do this so I wouldn't
forget about the ideas I come up with that can help me in creating the
script.