Full Backup Menu
I am half way finished with the Full Backup function. Although I am
still trying to figure out how to create the algorithm needed for this
to work I believe I can do it. There is an algorithm that I am trying
to do where the directory lists the files to backup and saves it
into an array. The array will be used to archive and compress the files
and directories all in one go. I hope that there is no problem
when naming the files.