Modified the function to archive and compress files.
The compressArchives () function is almost complete. It lists the files
and archives them using the for loop. I also wanted to add another feature
where when it creates the archives it saves them into a temporary directory.
From there, when it needs to transfer the files their destination it searches
for the compressed files inside the temporary directory and starts to
transfer them to their destination directory (remote or local).