Final modifications.
I rewrote some of the code so it can work with the rsync command. I chose
the options to this command because I thought it would work for the
occassion. This script still has room for improvement. I haven't tested it
yet but as soon as I do and it works as expected I will upload it to my
github account.